backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
motor
requests
python-dotenv
pinecone
pydantic
python-dotenv
pinecone_text
pydantic-settings
slowapi